refactor LoginForm
