+++ src/components/Login.tsx
@@ -1,15 +1,17 @@
-import { inject, observer } from 'mobx-react'
-import * as React from 'react'
-import { RouteComponentProps, withRouter } from 'react-router-dom'
+import React from 'react'
 import styled from 'styled-components'
-import { EventStore } from '../stores/EventStore'
-import { UserStore } from '../stores/UserStore'
-import { Header } from './Header'
-import { LoginForm } from './LoginForm/LoginForm'
+import { connect } from 'react-redux'
+import { RouteComponentProps, withRouter } from 'react-router-dom'
 
-interface Props extends RouteComponentProps {
-	userStore: UserStore
-	eventStore: EventStore
+import * as userActionCreator from '../ducks/user'
+
+import { Header } from './Header'
+import { LoginForm } from './LoginForm'
+
+import { Credential } from '../types/Credential'
+
+type Props = RouteComponentProps & {
+	login: (credential: Credential) => Promise<any>
 }
 
 const LoginFormWrapper = styled.div`
@@ -20,30 +22,27 @@ const LoginFormWrapper = styled.div`
 const Heading = styled.h2`
 	margin: 16px;
 `
-@inject('userStore', 'eventStore')
-@(withRouter as any)
-@observer
-export class Login extends React.Component<Props> {
-	public render(): React.ReactElement<any> {
-		const { userStore, eventStore, history } = this.props
-		return (
-			<div id="login">
-				<Header
-					history={history}
-					eventStore={eventStore}
-					userStore={userStore}
-				/>
-				<div className="login">
-					<LoginFormWrapper>
-						<Heading>
-							FOM: Fairy Online Manager
-							<br />
-							にログイン
-						</Heading>
-						<LoginForm userStore={userStore} history={history} />
-					</LoginFormWrapper>
-				</div>
+
+function LoginEntry({ history, login }: Props): React.ReactElement<{}> {
+	return (
+		<div id="login">
+			<Header history={history} />
+			<div className="login">
+				<LoginFormWrapper>
+					<Heading>
+						FOM: Fairy Online Manager
+						<br />
+						にログイン
+					</Heading>
+					<LoginForm login={login} history={history} />
+				</LoginFormWrapper>
 			</div>
-		)
-	}
+		</div>
+	)
 }
+
+export const Login = withRouter(
+	connect(null, {
+		login: userActionCreator.login,
+	})(LoginEntry)
+)
